Share common code for retrieving files.
This refactors getMyArDriveFilesFromPermaWeb, to better share common
logic for retrieving the list of files from Arweave between personal
(public/private) and shared drives.

With this change, the main logic is there only once in a helper
function, and it gets reused for all three types of drives rather
than being duplicated.