Batch database writes for retrieving files.
Batch (with a savepoint) all database writes for each drive when
retrieving the list of updated files from Arweave, and also when
fixing the file paths.

This has two benefits:  First it reduces disk writes (which would
otherwise be done for each single file) and thus improves performance,
and second it makes sure that in case of a crash or other failure
the updates are consistent (either a drive is fully updated to
the latest state, or it is not updated at all and will be retried later).