Function for retry / gateway switch.
There exists currently some ad-hoc code for retrying gateway requests
and switching out the gateway on error.  This introduces a general
method for doing so (with the actual request abstracted into an arbitrary
callback closure), which can be used in the future in place of all
those custom codes.

For now, we use it in getTransactionData and getPrivateTransactionCipherIV.