Query gateway directly for cipher IV.
When we need the cipher IV for a data transaction, query the
gateway directly with an /tx/txid URL to get the tags, rather
than using GQL.

Since we only look up by txid, there is no need for heavy querying
and instead the direct txid lookup is good enough.  This also helps
save GQL queries in case of rate limiting.