Query for data cipher IV when downloading a file.
We only need the data cipher IV when we actually download a private
file, not already when we sync it.  Thus we can move the query
to the download phase (and leave the field in the database empty);
this helps speed up the sync, especially if some folders are cloud-only.